+++ src/formlogin/Main.java
@@ -1,0 +1,9 @@
+package formlogin;
+
+public class Main {
+    public static void main(String[] args) {
+        // ! memanggil frame
+        new FrameLogin("Form Login");
+    }
+
+}
